Move hooks&constant in MapPage
