+++ src/components/Mission/MissionBar/MissionBar.js
@@ -6,7 +6,7 @@ import Box from '@material-ui/core/Box'
 import { makeStyles } from '@material-ui/core/styles'
 import Typography from '@material-ui/core/Typography'
 
-import { useMissionValue } from '../../../pages/MapPage/contexts/MissionContext'
+import { useMissionValue } from '../../../utils/contexts/MissionContext'
 import Flag from '../../Flag'
 
 const useStyles = makeStyles((theme) => ({
