+++ src/pages/MapPage/MapPage.js
@@ -11,10 +11,10 @@ import ProfileDialog from '../../components/ProfileDialog/ProfileDialog'
 import {
   MissionContextProvider,
   useMissionValue
-} from './contexts/MissionContext'
+} from '../../utils/contexts/MissionContext'
 import ReportHistory from '../../components/ReportHistory'
 import GuidePage from '../../components/GuidePage'
-import { useTagValue } from './contexts/TagContext'
+import { useTagValue } from '../../utils/contexts/TagContext'
 import useStep from '../../utils/hooks/useStep'
 import TagDetailDialog from '../../components/TagDetail/TagDetailDialog'
 import FilterFab from '../../components/Filter/FilterFab'
